refactor to use custom inclusion tag for stocks
