+++ articles/static/articles/style.css
@@ -50,30 +50,6 @@
     display: none;
     border: rgb(43, 38, 38);
 }
-.stock-header{
-    font-size: 22px;
-    margin-bottom: 4px;
-    color: cornflowerblue;
-    font-weight: 400;
-
-}
-.stock-container {
-    margin-bottom: 8px;
-}
-.stock-company {
-    font-weight: bold;
-    font-size: 18px;
-}
-.stock-info {
-    color: rgb(47, 90, 119);
-    font-weight: bold;
-}
-.stock-change-down {
-    color: red;
-}
-.stock-change-up {
-    color: green;
-}
 .search-results-header {    
     font-weight: bold;
     font-size: 18px;
