+++ articles/tests.py
@@ -48,7 +48,3 @@ class ContentApiTests(TestCase):
         self.assertEquals(article.uuid, main_article_uuid)
         self.assertEquals(article.byline, byline)
         self.assertEquals(article.image_url, image_url)
-
-class QuoteApiTests(TestCase):
-    def test_get_stocks_from_api(self):
-        quote_list = get_stocks_from_api()
