+++ stocks/static/stocks/stocks.js
